DE2/dashboard/requirements.txt
streamlit>=1.32
pandas>=2.1
plotly>=5.19
google-cloud-bigquery>=3.17
google-cloud-bigquery-storage>=2.24
pyarrow>=15.0
db-dtypes>=1.2
